ai_service/requirements.txt
fastapi
uvicorn
gunicorn
google-cloud-storage
google-cloud-aiplatform